Bring back Aeson number formatting as the default.
This default is actually quite sensible: not only does it reasonably switch
between fixed-decimal and scientific notation, it also discerns fractional
from integral numbers.

The new 'Generic' option will simply print _all_ numbers in floating-point
format, which is frankly a bit obnoxious seeing as JSON is a dynamic format
and uses 'Number' for numbers of any type.
E.g. 17.0 is not really an integer, is it?